refactor: merge variable statement checkers into single VariablesChecker
Continues the single-visitor migration. Merges four AST visitors that all
walked variable-related statements into one `VariablesChecker` in the new
`linter/rules/variable_statements.py`:

- misc.ConsistentAssignmentSignChecker
- misc.EmptyVariableChecker
- misc.NonLocalVariableChecker
- misc.MissingVariableTypeChecker

The new checker owns 10 rules spanning three categories (variables, typing,
misc) and the decision logic moved into `check()` methods on the rules
themselves, matching the pattern used by the previously merged checkers.

The `empty-variable` `variable_source` parameter previously gated *traversal*
(the old checker only descended into a section when that source was enabled).
Since the merged visitor must always descend for the other rules, the gating
moved to a `check_empty_variable` flag consulted in `visit_Variable` /
`visit_Var`. `Variable` statements only appear in `*** Variables ***` and
`VAR` only in keyword/test bodies, so this is behaviourally equivalent.

Checker count: 33 -> 30. No user-visible change: verified byte-identical
output against the previous revision for a full `--select ALL` scan of the
test corpus (189296 lines), per-rule runs for all 10 rules, 8 configured
variants of the gating parameters, and a hand-written edge-case suite.
